split the symbolic function factories into their own header files
