+++ src/Core/Utility/CMakeLists.txt
@@ -42,7 +42,8 @@ SET(CYCLUS_CORE_SRC ${CYCLUS_CORE_SRC}
   ${CMAKE_CURRENT_SOURCE_DIR}/Logger.cpp 
   ${CMAKE_CURRENT_SOURCE_DIR}/MassTable.cpp 
   ${CMAKE_CURRENT_SOURCE_DIR}/RecipeLogger.cpp 
-  ${CMAKE_CURRENT_SOURCE_DIR}/SupplyDemandManager.cpp 
+  ${CMAKE_CURRENT_SOURCE_DIR}/SupplyDemandManager.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/SymbolicFunctionFactories.cpp  
   ${CMAKE_CURRENT_SOURCE_DIR}/SymbolicFunctions.cpp 
   ${CMAKE_CURRENT_SOURCE_DIR}/Table.cpp 
   ${CMAKE_CURRENT_SOURCE_DIR}/Timer.cpp 
@@ -62,6 +63,7 @@ INSTALL(FILES
   RecipeLogger.h
   SupplyDemand.h
   SupplyDemandManager.h
+  SymbolicFunctionFactories.h
   SymbolicFunctions.h
   Table.h
   Timer.h
